Angular and Promise Part 1
A promise is a mechanism that supports asynchronous computation.
In using a promise, the promise provides a mechanism or a proxy for a value that is not available at the moment that you ask for it.
So a promise will return a promise to you, saying that when the result becomes available, it'll come back and deliver the results to you.
So a promise returns a proxy object to you, and you hold on to the proxy object.

It may be also immediately available, in which case the promise will return the value immediately.
But in case it needs to return a value at a future date, then it gives you the promise that that value when it becomes available will be returned to you.

* Our first step in this exercise is to reconfigure all the services to return promises. So each of the metrics within our service, instead of returning the value immediately, it'll return promises.
* Then, Consume the promises within your components and handle the resolution or rejection of the promise