Angular and RxJS Part 1
* Make use of observables within your Angular application (convert the services to use Observables in place of promises)
* Obtain data through subscribing to the observables within your components.

Reactive programming is based around observer pattern and also an iterated pattern.
in reactive programming, we are interested in data flows. How data flows through your application.
Any data flow can be treated as a stream. So this data flow can propagate changes throughout your application.
(at intermediary stages you can even take one particular stream, modify it and create another stream going out from there. So that is what an observable pattern enables us to drew.)

RxJS is a JavaScript implementation of the observable pattern and also support for reactive programming.
RxJS is a library that enables you to do asynchronous programming.
So, when you are wanting to wait for an asynchronous event, it will be similar to subscribing to an observable.
So when that observable emits the value, you are ready to catch the value and then act accordingly. You do know when that emission is going to take place.
So that is the asynchronous part coming into play with the use of RxJS.
Angular leverages RxJS for its support for reactive programming.
Forms themselves become observables in Angular. And any changes in the values of any of the form elements, can be outputted, and observed by subscribing to the observable that the Angular framework provides for us.