Angular and REST (Representational State Transfer) : Restangular
* Enable your Angular application to use Ng2-Restangular
npm install --save ngx-restangular@1.0.13
* Make use of Ngx-Restangular to access a server that supports a RESTful API.

Web services are a way of designing systems to support interoperability among systems that are connected over a network,
This means is that you provide a standardized way for two machines that are connected to the internet to be able to communicate with each other.
Two common approaches that are used for supporting web services are
SOAP (using XML as the format for the messages being exchanged between the two endpoints.)
and REST (Using JSON, simpler compared to SOAP and hence REST has found a lot wider deployment )

, with AngularJS, you know that there was a built-in service called resource that facilitated access to REST API-based servers.
But with the new Angular framework, no such service exists pre-built into the Angular framework itself.
Instead, Angular expects you to resort to using the HTTP module for this purpose.
Fortunately, third-party libraries are available to fill in this void, and one such library is Restangular

Restangular is an Angular 2+ service designed for the new Angular framework that simplifies the common GET, PUT, POST, and UPDATE methods writing the minimum amount of client code