Angular Animations Part 1
* Define new animations using the support available in the Angular framework
* Apply the animations to the views within your components

Animations are not necessary or required but they add that little spice to your angular application,
if you're keep that view in your mind then you will approach animation from a completely different perspective.
You need to look at it from the user experience, or the user interface design perspective, then you begin to see that animations have a meaning and a purpose for their use in an application.
( Things like, for example, when you click on a button, the ripple effects that you see on the button and the subtle enlarging, and shrinking of different parts of your view, and so on. )
The animations themselves are built around states and transitions among states in the Angular framework

json-server --watch db.json -d 2000