Angular Animations Part 2
* Define new animations using the support available in the Angular framework
* Apply the animations to the views within your components

we will add in an animation that will be triggered when the route changes occur within our single page application.
We will also add in an animation which enables a view to be displayed when the data is fetched from the server side.
So when you are dismissing the spinner in your view and then rendering the view, we want the view to be projected on
the screen rather than simply just appearing all of a sudden.